[tests] New tests for Command Line formatting
Move generic attributes formatting in a dedicated unit test.
